refactor(ops): address PR review comments for metrics (#12)
- Extract metrics code to cli/metrics.rs (serve.rs: 865 -> 720 lines)
- Replace .expect() with .unwrap_or_else() in /metrics handler
- Move bytes_received_total increment before persistence in bulk_put
- Fix stale bucket clone to use &str then .to_owned()
